Dividing some code by modules... players, chars, logging
