+++ tools.js
@@ -1,0 +1,1 @@
+export const getRandom = (range) => Math.ceil(Math.random() * range);
